+++ demo/routes/Notification/index.js
@@ -1,0 +1,8 @@
+export default {
+  path: 'notification',
+  getComponent(nextState, cb) {
+    require.ensure([], (require) => {
+      cb(null, require('./components/Notification').default);
+    });
+  }
+};